Improved rendering of images, exctracted parseData and renderHighlights
